Fix: Parts fiscales crash (int->decimal) + implement PAS calculation from quotient familial
- PayrollRequestDto: Parts int -> decimal (1.5, 2.5 etc. now deserialize correctly)
- PayrollService.Simulate(): if RetenuePct=0 and Parts>0, compute PAS rate via
  ComputePasTaux() using quotient familial (netAnnuel / parts) and barème 2026
- Add ComputePasTaux(): barème PAS 2026 sur revenu par part (0/2/7.5/14/22/30/41%)
- PayrollResponseDto: add Parts and NetImposable fields for frontend display

+++ src/Application/DTOs/PayrollRequestDto.cs
@@ -5,7 +5,7 @@ namespace MegaSimulator.Application.DTOs
         public decimal Brut { get; set; }
         public string Statut { get; set; } = "non-cadre";
         public decimal BrutAnnuel { get; set; }
-        public int Parts { get; set; } = 1;
+        public decimal Parts { get; set; } = 1m;
         public decimal RevenusAnnexes { get; set; }
         public decimal Primes { get; set; }
         public decimal RetenuePct { get; set; }
